lib/syscall_shim: Compile-time mapping of variadic system calls
Variadic system calls are system calls with a variable number of
arguments. Some of them become optional or are only needed depending on
a given flag or a command number.
This commit introduces the support for statically mapping such
system calls at compile time. For this purpose, the
`uk_syscall_(r_)static()` macros are evaluating the number of given
arguments and will automatically auto-complete missing arguments with
zeros or remove additional arguments that aren't handled by the
particular system call implementation.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>