plat/pci_ecam: Fix FDT node lacking address-cells
The `gen_pci_irq_parse` function fails when an interrupt parent node is missing the #address-cells property. Even though the devicetree spec states that this property shall be explicitly defined, the RISC-V DTB generated by QEMU doesn't include this property for the interrupt controller. However, if the parent unit address is missing form the DTB, we can assume that the parent node's address-size property is equal to 0.

Signed-off-by: Eduard Vintilă <eduard.vintila47@gmail.com>