plat/*: Add riscv64 implementation of the paging API
This commit implements the architecture-dependent parts of the
paging API for the riscv64 Sv39 mode and extends the platform setup code to
use this API, thus replacing the old paging implementation written for riscv64.

For consistency, some code is taken directly from the x86 implementation.
For example, we use the same frame
allocator (ukfallocbuddy), and physical memory ranges are added to this
allocator in exactly the same manner.

We also have an area in the virtual address space starting from 256GiB
which is used to map the physical memory.
For the moment, we identity map the MMIO/PCI-e space (0x0 - 0x7fffffff) so
we can directly use the physical
addresses parsed from the DTB.
However, maybe in the future a better idea would be to do MMIO through the
direct map and abandon this identity mapping.

The platform setup code which constructs a dynamic page table based
on the static page table found at plat/kvm/riscv/pagetable64.S is simillar
to the x86 one, except we explicitly map
sections of the kernel image with their appropriate attributes (.text as read-
executable only, for example).

Signed-off-by: Eduard Vintilă <eduard.vintila47@gmail.com>