nolibc: Add net/if.h from Musl
LWIP has been updated to work with Musl. Headers and source code files
from LWIP have been removed since they are part of Musl.

In order to keep nolibc working with LWIP, previous items part of LWIP
now have to be part of nolibc.

This commit adds the net/if.h header from Musl. It was previously part
of LWIP, but has now been removed to make it compatible with Musl.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>