plat/ofw: Fix fdt functions
This commit fixes issues found in two functions: `fdt_translate_one` and
`fdt_node_offset_idx_by_compatible_list`:

`fdt_translate_one`:
When a node is missing the "ranges" property, offset should be set to 0.
The current implementation sets offset to the address to be translated,
which ignores the fact that the fdt_default_translate function is later
called. This function adds the offset to the same address,
resulting in doubling the effective address, when in fact a 1-to-1
mapping is necessary.

`fdt_node_offset_idx_by_compatible_list`:
The current implementation returns as soon as a node that matches a
compatible string is found. However, we should iterate through all of
the compatible list to make sure that we return the node with the least
offset value such that future searches will not skip any matching nodes.

Signed-off-by: Eduard Vintilă <eduard.vintila47@gmail.com>